+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.6.2
+* General:
+    * Additional output file selection in Export Option Menu
 * Internal:
     * removed `use strict` due to modul type
     * added class comments